oxford_flowers_102/requirements.txt
# Oxford Flowers 102 PyTorch Classifier Requirements
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
tqdm>=4.65.0  # For feature extraction progress bars
